requirements.txt
flask
line-bot-sdk<3.0.0
requests
